Bug, Guardado y particulas
Se agregaron el sistema de particulas para arboles y piedras, se cambio el sistema de objetos y se agrego el guardado automatico del inventario

+++ Assets/Mis Cosas/Scripits/UI/Scr_ObjetoEnMano.cs
@@ -171,7 +171,7 @@ public class Scr_ObjetoEnMano : MonoBehaviour
 
             Casilla.GetComponent<Scr_CasillaInventario>().FormaConHermanas = Forma;
 
-            Inventario.Cantidades[(int)Casilla.GetComponent<Scr_CasillaInventario>().Numero] += Cantidad;
+            Inventario.Cantidades[(int)Casilla.GetComponent<Scr_CasillaInventario>().Numero] += (int)Cantidad;
         }
     }
